toolkit/content/tests/widgets/test_videocontrols_iframe_fullscreen.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Video controls test - iframe</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="head.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>

<div id="content">
<iframe id="ifr1"></iframe>
<iframe id="ifr2" allowfullscreen></iframe>
<iframe id="ifr1" allow="fullscreen 'none'"></iframe>
</div>

<pre id="test">
<script clas="testbody" type="application/javascript">
  SimpleTest.waitForExplicitFinish();

  const testCases = [];

  function checkIframeFullscreenAvailable(ifr) {
    let video;

    return () => new Promise(resolve => {
      ifr.srcdoc = `<video id="video" controls preload="auto"></video>`;
      ifr.addEventListener("load", resolve);
    }).then(() => new Promise(resolve => {
      video = ifr.contentDocument.getElementById("video");
      video.src = "seek_with_sound.ogg";
      video.addEventListener("loadedmetadata", resolve);
    })).then(() => new Promise(resolve => {
      const available = video.ownerDocument.fullscreenEnabled;
      const controlBar = getElementWithinVideo(video, "controlBar");

      is(controlBar.getAttribute("fullscreen-unavailable") == "true", !available, "The controlbar should have an attribute marking whether fullscreen is available that corresponds to if the iframe has the allowfullscreen attribute.");
      resolve();
    }));
  }

  function start() {
    testCases.reduce((promise, task) => promise.then(task), Promise.resolve());
  }

  function load() {
    SpecialPowers.pushPrefEnv({"set": [["media.cache_size", 40000]]}, start);
  }

  for (let iframe of document.querySelectorAll("iframe"))
    testCases.push(checkIframeFullscreenAvailable(iframe));
  testCases.push(SimpleTest.finish);

  window.addEventListener("load", load);
</script>
</pre>
</body>
</html>
